feat(api): refatorar api para consumir pacotes compartilhados (#9)
- Remove src/app.ts (inline server creation in index.ts)
- Remove src/db/index.ts (db now lives in @spec-driven/db)
- Flatten index.ts to use createServer and createFastifyLoggerConfig directly
- Add db:generate, db:migrate, db:studio scripts delegating to @spec-driven/db
