+++ apps/api/package.json
@@ -7,7 +7,10 @@
     "dev": "tsx watch src/index.ts",
     "build": "tsup",
     "start": "node dist/index.js",
-    "check-types": "tsc --noEmit"
+    "check-types": "tsc --noEmit",
+    "db:generate": "pnpm --filter @spec-driven/db db:generate",
+    "db:migrate": "pnpm --filter @spec-driven/db db:migrate",
+    "db:studio": "pnpm --filter @spec-driven/db db:studio"
   },
   "dependencies": {
     "@spec-driven/db": "workspace:*",
