--- apps/api/src/app.ts
@@ -1,14 +1,0 @@
-import { createServer } from "@spec-driven/http"
-import { createFastifyLoggerConfig } from "@spec-driven/logger"
-
-export function buildApp() {
-  const server = createServer({
-    logger: createFastifyLoggerConfig({
-      name: "api",
-      level: process.env.LOG_LEVEL ?? "info",
-    }),
-    cors: { origin: true },
-  })
-
-  return server
-}
