--- apps/api/src/db/index.ts
@@ -1,7 +1,0 @@
-import { createDatabase } from "@spec-driven/db"
-
-if (!process.env.DATABASE_URL) {
-  throw new Error("DATABASE_URL environment variable is required")
-}
-
-export const db = createDatabase(process.env.DATABASE_URL)
