+++ apps/api/src/index.ts
@@ -1,13 +1,22 @@
-import { buildApp } from "./app.js"
+import { createServer } from "@spec-driven/http"
+import { createFastifyLoggerConfig } from "@spec-driven/logger"
 
-const PORT = Number(process.env.PORT) || 3001
-const HOST = process.env.HOST ?? "0.0.0.0"
+const server = createServer({
+  logger: createFastifyLoggerConfig({
+    name: "api",
+    level: process.env.LOG_LEVEL ?? "info",
+  }),
+  cors: { origin: true },
+})
 
-const app = buildApp()
+// register routes...
 
 try {
-  await app.start({ port: PORT, host: HOST })
+  await server.start({
+    port: Number(process.env.PORT) || 3001,
+    host: process.env.HOST ?? "0.0.0.0",
+  })
 } catch (err) {
-  app.log.error(err)
+  server.log.error(err)
   process.exit(1)
 }
